Extract subcommands
