+++ bin/cml.test.js
@@ -8,30 +8,26 @@ describe('command-line interface tests', () => {
       "cml.js <command>
 
       Commands:
-        cml.js ci                                 Fixes specific CI setups
-        cml.js pr <glob path...>                  Create a pull request with the
-                                                  specified files
-        cml.js rerun-workflow                     Reruns a workflow given the jobId or
-                                                  workflow Id
-        cml.js runner                             Launch and register a self-hosted
-                                                  runner
-        cml.js send-comment <markdown file>       Comment on a commit
-        cml.js send-github-check <markdown file>  Create a check report
-        cml.js tensorboard-dev                    Get a tensorboard link
+        cml.js check        Manage continuous integration checks
+        cml.js pr           Manage pull requests
+        cml.js report       Manage reports
+        cml.js repository   Manage repository settings
+        cml.js runner       Manage continuous integration self-hosted runners
+        cml.js tensorboard  Manage tensorboard.dev agents
+        cml.js workflow     Manage continuous integration workflows
 
       Options:
         --help     Show help                                                 [boolean]
         --version  Show version number                                       [boolean]
         --log      Maximum log level
                 [string] [choices: \\"error\\", \\"warn\\", \\"info\\", \\"debug\\"] [default: \\"info\\"]
-        --driver   Platform where the repository is hosted. If not specified, it will
-                   be inferred from the environment
+        --driver   Forge where the repository is hosted. If not specified, it will be
+                   inferred from the environment
                                    [string] [choices: \\"github\\", \\"gitlab\\", \\"bitbucket\\"]
-        --repo     Repository to be used for registering the runner. If not specified,
-                   it will be inferred from the environment                   [string]
-        --token    Personal access token to register a self-hosted runner on the
-                   repository. If not specified, it will be inferred from the
-                   environment                                                [string]"
+        --repo     Repository. If not specified, it will be inferred from the
+                   environment                                                [string]
+        --token    Personal access token. If not specified, it will be inferred from
+                   the environment                                            [string]"
     `);
   });
 });
