+++ bin/cml/attachment/publish.test.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const { exec } = require('../../src/utils');
+const { exec } = require('../../../src/utils');
 
 describe('CML e2e', () => {
   test('cml publish --help', async () => {
@@ -13,15 +13,13 @@ describe('CML e2e', () => {
             --version       Show version number                              [boolean]
             --log           Maximum log level
                 [string] [choices: \\"error\\", \\"warn\\", \\"info\\", \\"debug\\"] [default: \\"info\\"]
-            --driver        Platform where the repository is hosted. If not specified,
-                            it will be inferred from the environment
+            --driver        Forge where the repository is hosted. If not specified, it
+                            will be inferred from the environment
                                    [string] [choices: \\"github\\", \\"gitlab\\", \\"bitbucket\\"]
-            --repo          Repository to be used for registering the runner. If not
-                            specified, it will be inferred from the environment
-                                                                              [string]
-            --token         Personal access token to register a self-hosted runner on
-                            the repository. If not specified, it will be inferred from
-                            the environment                                   [string]
+            --repo          Repository. If not specified, it will be inferred from the
+                            environment                                       [string]
+            --token         Personal access token. If not specified, it will be
+                            inferred from the environment                     [string]
             --md            Output in markdown format [title || name](url).  [boolean]
         -t, --title         Markdown title [title](url) or ![](url title).    [string]
             --native        Uses driver's native capabilities to upload assets instead
