+++ bin/cml/report.js
@@ -1,0 +1,8 @@
+exports.command = 'report';
+exports.description = 'Manage reports';
+exports.builder = (yargs) =>
+  yargs
+    .commandDir('./report', { exclude: /\.test\.js$/ })
+    .recommendCommands()
+    .demandCommand()
+    .strict();
