+++ bin/cml/runner/start.test.js
@@ -1,8 +1,8 @@
 jest.setTimeout(2000000);
 
 const isIp = require('is-ip');
-const { CML } = require('../..//src/cml');
-const { exec, sshConnection, randid, sleep } = require('../../src/utils');
+const { CML } = require('../../../src/cml');
+const { exec, sshConnection, randid, sleep } = require('../../../src/utils');
 
 const IDLE_TIMEOUT = 15;
 const {
@@ -56,26 +56,26 @@ describe('CML e2e', () => {
     expect(output).toMatchInlineSnapshot(`
       "cml.js runner
 
-      Launch and register a self-hosted runner
+      Manage continuous integration self-hosted runners
+
+      Commands:
+        cml.js runner start  Start and register a self-hosted runner
 
       Options:
         --help                                    Show help                  [boolean]
         --version                                 Show version number        [boolean]
         --log                                     Maximum log level
                 [string] [choices: \\"error\\", \\"warn\\", \\"info\\", \\"debug\\"] [default: \\"info\\"]
-        --driver                                  Platform where the repository is
+        --driver                                  Forge where the repository is
                                                   hosted. If not specified, it will be
                                                   inferred from the environment
                                    [string] [choices: \\"github\\", \\"gitlab\\", \\"bitbucket\\"]
-        --repo                                    Repository to be used for
-                                                  registering the runner. If not
-                                                  specified, it will be inferred from
-                                                  the environment             [string]
-        --token                                   Personal access token to register a
-                                                  self-hosted runner on the
-                                                  repository. If not specified, it
+        --repo                                    Repository. If not specified, it
                                                   will be inferred from the
                                                   environment                 [string]
+        --token                                   Personal access token. If not
+                                                  specified, it will be inferred from
+                                                  the environment             [string]
         --labels                                  One or more user-defined labels for
                                                   this runner (delimited with commas)
                                                              [string] [default: \\"cml\\"]
