+++ bin/cml/tensorboard/start.test.js
@@ -1,7 +1,7 @@
 const fs = require('fs').promises;
 const tempy = require('tempy');
-const { exec, isProcRunning, sleep } = require('../../src/utils');
-const { tbLink } = require('./tensorboard-dev');
+const { exec, isProcRunning, sleep } = require('../../../src/utils');
+const { tbLink } = require('./start');
 
 const CREDENTIALS =
   '{"refresh_token": "1//03FiVnGk2xhnNCgYIARAAGAMSNwF-L9IrPH8FOOVWEYUihFDToqxyLArxfnbKFmxEfhzys_KYVVzBisYlAy225w4HaX3ais5TV_Q", "token_uri": "https://oauth2.googleapis.com/token", "client_id": "373649185512-8v619h5kft38l4456nm2dj4ubeqsrvh6.apps.googleusercontent.com", "client_secret": "pOyAuU2yq2arsM98Bw5hwYtr", "scopes": ["openid", "https://www.googleapis.com/auth/userinfo.email"], "type": "authorized_user"}';
@@ -57,22 +57,18 @@ describe('CML e2e', () => {
     expect(output).toMatchInlineSnapshot(`
       "cml.js tensorboard-dev
 
-      Get a tensorboard link
-
       Options:
             --help          Show help                                        [boolean]
             --version       Show version number                              [boolean]
             --log           Maximum log level
                 [string] [choices: \\"error\\", \\"warn\\", \\"info\\", \\"debug\\"] [default: \\"info\\"]
-            --driver        Platform where the repository is hosted. If not specified,
-                            it will be inferred from the environment
+            --driver        Forge where the repository is hosted. If not specified, it
+                            will be inferred from the environment
                                    [string] [choices: \\"github\\", \\"gitlab\\", \\"bitbucket\\"]
-            --repo          Repository to be used for registering the runner. If not
-                            specified, it will be inferred from the environment
-                                                                              [string]
-            --token         Personal access token to register a self-hosted runner on
-                            the repository. If not specified, it will be inferred from
-                            the environment                                   [string]
+            --repo          Repository. If not specified, it will be inferred from the
+                            environment                                       [string]
+            --token         Personal access token. If not specified, it will be
+                            inferred from the environment                     [string]
         -c, --credentials   TB credentials as json. Usually found at
                             ~/.config/tensorboard/credentials/uploader-creds.json. If
                             not specified will look for the json at the env variable
