+++ bin/cml/workflow.js
@@ -1,0 +1,8 @@
+exports.command = 'workflow';
+exports.description = 'Manage continuous integration workflows';
+exports.builder = (yargs) =>
+  yargs
+    .commandDir('./workflow', { exclude: /\.test\.js$/ })
+    .recommendCommands()
+    .demandCommand()
+    .strict();
