+++ bin/legacy/rerun-workflow.js
@@ -1,0 +1,6 @@
+const { builder, handler } = require('../cml/workflow/restart');
+
+exports.command = 'rerun-workflow';
+exports.description = false;
+exports.handler = handler;
+exports.builder = builder;
