+++ bin/legacy/tensorboard-dev.js
@@ -1,0 +1,6 @@
+const { builder, handler } = require('../cml/tensorboard/start');
+
+exports.command = 'tensorboard-dev';
+exports.description = false;
+exports.handler = handler;
+exports.builder = builder;
